fix(virtq): move log tests to integration tests
